Move texture.h to common/
