+++ src/common/texture.h
